+++ packages/vscode-extension-serverless-workflow-editor/src/extension/extension.ts
@@ -70,6 +70,16 @@ export async function activate(context: vscode.ExtensionContext) {
   const swEnvelopeType = "swf";
   const baseEnvelopePath = "dist/webview/editors/serverless-workflow";
 
+  const diagramEnvelopeMappingConfig = configuration.isKogitoServerlessWorkflowVisualizationPreviewEnabled()
+    ? {
+        resourcesPathPrefix: baseEnvelopePath + "/diagram",
+        envelopePath: baseEnvelopePath + "/serverless-workflow-diagram-editor-envelope.js",
+      }
+    : {
+        resourcesPathPrefix: baseEnvelopePath,
+        envelopePath: baseEnvelopePath + "/serverless-workflow-mermaid-viewer-envelope.js",
+      };
+
   const kieToolsEditorStore = await KieToolsVsCodeExtensions.startExtension({
     editorDocumentType: "text",
     extensionName: "kie-group.vscode-extension-serverless-workflow-editor",
@@ -81,8 +91,8 @@ export async function activate(context: vscode.ExtensionContext) {
       new EnvelopeMapping(
         swEnvelopeType,
         "**/*.sw.json",
-        baseEnvelopePath + "/serverless-workflow-diagram-editor-envelope.js",
-        baseEnvelopePath + "/diagram"
+        diagramEnvelopeMappingConfig.envelopePath,
+        diagramEnvelopeMappingConfig.resourcesPathPrefix
       ),
       new EnvelopeMapping(
         swEnvelopeType,